+++ src/kernel/vect_pos.c
@@ -11,8 +11,8 @@ void pos_robot_to_table(struct vect_pos *pos_robot, struct vect_pos *pos_in, str
 	pos_out->x = pos_robot->ca * pos_in->x - pos_robot->sa * pos_in->y + pos_robot->x;
 	pos_out->y = pos_robot->sa * pos_in->x + pos_robot->ca * pos_in->y + pos_robot->y;
 	pos_out->alpha = pos_in->alpha + pos_robot->alpha;
-	pos_out->ca = cos(pos_out->alpha);
-	pos_out->sa = sin(pos_out->alpha);
+	pos_out->ca = cosf(pos_out->alpha);
+	pos_out->sa = sinf(pos_out->alpha);
 }
 
 //! changement de repere du repère robot au repere table en fonction de la position du robot
@@ -21,8 +21,8 @@ void pos_table_to_robot(struct vect_pos *pos_robot, struct vect_pos *pos_in, str
 	pos_out->x = pos_robot->ca * pos_in->x + pos_robot->sa * pos_in->y - pos_robot->x;
 	pos_out->y = - pos_robot->sa * pos_in->x + pos_robot->ca * pos_in->y - pos_robot->y;
 	pos_out->alpha = pos_in->alpha - pos_robot->alpha;
-	pos_out->ca = cos(pos_out->alpha);
-	pos_out->sa = sin(pos_out->alpha);
+	pos_out->ca = cosf(pos_out->alpha);
+	pos_out->sa = sinf(pos_out->alpha);
 }
 
 //! changement de repere du repère hokuyo au repere table en fonction de la position du robot
@@ -33,8 +33,8 @@ void pos_hokuyo_to_table(struct vect_pos *pos_robot, struct vect_pos *pos_in, st
 	pos_out->y = pos_robot->sa * x + pos_robot->ca * pos_in->y + pos_robot->y;
 // on n'a pas besoin de l'angle
 	pos_out->alpha = 0; //pos_in->alpha + pos_robot->alpha;
-	pos_out->ca = 1; //cos(pos_out->alpha);
-	pos_out->sa = 0; //sin(pos_out->alpha);
+	pos_out->ca = 1; //cosf(pos_out->alpha);
+	pos_out->sa = 0; //sinf(pos_out->alpha);
 }
 
 float norm2_square(struct vect_pos *pos)
